# very small (but effective) performance improvement for VectorNumeric values