@levens 2016-11-16
Hi Michi,

Nice improvement. Can you please also bump the version number in `__init__.py`.

Cheers,