@rdemaria 2016-11-16
Does the improvement work also on the other get... methods?