@levens 2016-11-16
Possibly on `MATRIXNUMERIC` as well?