@michi42 2016-11-16
Well, it only makes sense for the multi-valued types - iterating over the time dimension is still required, but it replaces the implicit iteration over the inner dimension of VECTORNUMERIC values by the more efficient jpype slicing.

It might work for VECTORSTRING and MATRIXNUMERIC I presume, but I'd need to find some variable which uses these data types to verify that.