@michi42 2016-11-16
I just tried with two variables (LHC.BOFSU:BPM_NAMES_H for VectorString and LHC.BLMS:LOSS_FAST for MatrixNumeric), and for these the slicing does not improve the performance - for the string case it even looks slightly worse.

From looking at the profile results, I believe this comes from the fact that in both of these cases, we're not dealing with primitive java arrays (e.g. double[]), but rather with array of objects...