@chernals 2016-11-16
Nicely done. I think there are a few tricks in the multiprocessing branch (not related to that) that went a bit in that direction, but it's nicer here.