@levens 2016-11-16
BTW, we already play this trick in PyJAPC :-)