@levens 2016-11-17
FYI, the bdisoft python distribution has been updated.